pyproject.toml
[tool.poetry]
name = "test-project-cities"
version = "0.1.0"
description = "Test task for interview"
authors = ["Madiyar <madiarashkenovm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
